+++ notifications/client/client.go
@@ -5,9 +5,9 @@ package client
 import (
 	os "os"
 
-	core "github.com/KardFinancial/kard-go-sdk/v15/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v15/internal"
-	subscriptions "github.com/KardFinancial/kard-go-sdk/v15/notifications/subscriptions"
+	core "github.com/KardFinancial/kard-go-sdk/v16/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v16/internal"
+	subscriptions "github.com/KardFinancial/kard-go-sdk/v16/notifications/subscriptions"
 )
 
 type Client struct {