+++ notifications/subscriptions/notifications_subscriptions_test/notifications_subscriptions_test.go
@@ -10,10 +10,10 @@ import (
 	os "os"
 	testing "testing"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v15"
-	client "github.com/KardFinancial/kard-go-sdk/v15/client"
-	notifications "github.com/KardFinancial/kard-go-sdk/v15/notifications"
-	option "github.com/KardFinancial/kard-go-sdk/v15/option"
+	kard "github.com/KardFinancial/kard-go-sdk/v16"
+	client "github.com/KardFinancial/kard-go-sdk/v16/client"
+	notifications "github.com/KardFinancial/kard-go-sdk/v16/notifications"
+	option "github.com/KardFinancial/kard-go-sdk/v16/option"
 	require "github.com/stretchr/testify/require"
 )