+++ organizations/children/client.go
@@ -6,11 +6,11 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v15"
-	core "github.com/KardFinancial/kard-go-sdk/v15/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v15/internal"
-	option "github.com/KardFinancial/kard-go-sdk/v15/option"
-	organizations "github.com/KardFinancial/kard-go-sdk/v15/organizations"
+	kard "github.com/KardFinancial/kard-go-sdk/v16"
+	core "github.com/KardFinancial/kard-go-sdk/v16/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v16/internal"
+	option "github.com/KardFinancial/kard-go-sdk/v16/option"
+	organizations "github.com/KardFinancial/kard-go-sdk/v16/organizations"
 )
 
 type Client struct {