+++ organizations/client/client.go
@@ -6,13 +6,13 @@ import (
 	context "context"
 	os "os"
 
-	kard "github.com/KardFinancial/kard-go-sdk/v15"
-	core "github.com/KardFinancial/kard-go-sdk/v15/core"
-	internal "github.com/KardFinancial/kard-go-sdk/v15/internal"
-	option "github.com/KardFinancial/kard-go-sdk/v15/option"
-	children "github.com/KardFinancial/kard-go-sdk/v15/organizations/children"
-	contentstrategies "github.com/KardFinancial/kard-go-sdk/v15/organizations/contentstrategies"
-	placements "github.com/KardFinancial/kard-go-sdk/v15/organizations/placements"
+	kard "github.com/KardFinancial/kard-go-sdk/v16"
+	core "github.com/KardFinancial/kard-go-sdk/v16/core"
+	internal "github.com/KardFinancial/kard-go-sdk/v16/internal"
+	option "github.com/KardFinancial/kard-go-sdk/v16/option"
+	children "github.com/KardFinancial/kard-go-sdk/v16/organizations/children"
+	contentstrategies "github.com/KardFinancial/kard-go-sdk/v16/organizations/contentstrategies"
+	placements "github.com/KardFinancial/kard-go-sdk/v16/organizations/placements"
 )
 
 type Client struct {